# internal/server/payments.service.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for PostCallbacksTypeParamsType.
const (
	Json PostCallbacksTypeParamsType = "json"
	Xml  PostCallbacksTypeParamsType = "xml"
)

// Defines values for PostCallbacksTypeJSONBodyStatus.
const (
	Failure PostCallbacksTypeJSONBodyStatus = "failure"
	Pending PostCallbacksTypeJSONBodyStatus = "pending"
	Success PostCallbacksTypeJSONBodyStatus = "success"
)

// Defines values for PostTransactionsGatewayDepositParamsGateway.
const (
	GatewayA PostTransactionsGatewayDepositParamsGateway = "gatewayA"
	GatewayB PostTransactionsGatewayDepositParamsGateway = "gatewayB"
)

// PostCallbacksTypeJSONBody defines parameters for PostCallbacksType.
type PostCallbacksTypeJSONBody struct {
	OrderId       *openapi_types.UUID              `json:"order_id,omitempty"`
	Status        *PostCallbacksTypeJSONBodyStatus `json:"status,omitempty"`
	TransactionId *string                          `json:"transactionId,omitempty"`
}

// PostCallbacksTypeParamsType defines parameters for PostCallbacksType.
type PostCallbacksTypeParamsType string

// PostCallbacksTypeJSONBodyStatus defines parameters for PostCallbacksType.
type PostCallbacksTypeJSONBodyStatus string

// PostTransactionsGatewayDepositJSONBody defines parameters for PostTransactionsGatewayDeposit.
type PostTransactionsGatewayDepositJSONBody struct {
	Amount   float32 `json:"amount"`
	Currency string  `json:"currency"`
	OrderId  string  `json:"order_id"`
}

// PostTransactionsGatewayDepositParamsGateway defines parameters for PostTransactionsGatewayDeposit.
type PostTransactionsGatewayDepositParamsGateway string

// PostCallbacksTypeJSONRequestBody defines body for PostCallbacksType for application/json ContentType.
type PostCallbacksTypeJSONRequestBody PostCallbacksTypeJSONBody

// PostTransactionsGatewayDepositJSONRequestBody defines body for PostTransactionsGatewayDeposit for application/json ContentType.
type PostTransactionsGatewayDepositJSONRequestBody PostTransactionsGatewayDepositJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Universal callback handler
	// (POST /callbacks/{type})
	PostCallbacksType(ctx echo.Context, pType PostCallbacksTypeParamsType) error
	// Handle deposit
	// (POST /transactions/{gateway}/deposit)
	PostTransactionsGatewayDeposit(ctx echo.Context, gateway PostTransactionsGatewayDepositParamsGateway) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostCallbacksType converts echo context to params.
func (w *ServerInterfaceWrapper) PostCallbacksType(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "type" -------------
	var pType PostCallbacksTypeParamsType

	err = runtime.BindStyledParameterWithOptions("simple", "type", ctx.Param("type"), &pType, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter type: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostCallbacksType(ctx, pType)
	return err
}

// PostTransactionsGatewayDeposit converts echo context to params.
func (w *ServerInterfaceWrapper) PostTransactionsGatewayDeposit(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "gateway" -------------
	var gateway PostTransactionsGatewayDepositParamsGateway

	err = runtime.BindStyledParameterWithOptions("simple", "gateway", ctx.Param("gateway"), &gateway, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter gateway: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTransactionsGatewayDeposit(ctx, gateway)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/callbacks/:type", wrapper.PostCallbacksType)
	router.POST(baseURL+"/transactions/:gateway/deposit", wrapper.PostTransactionsGatewayDeposit)

}